+++ test.py
@@ -74,6 +74,11 @@ def test_messages():
         assert_true(False, 'should have raised an exception')
     except MultipleInvalid as e:
         assert_equal(1, len(e.errors))
+    try:
+        schema.to_native([{'aDict':{}}])
+        assert_true(False, 'should have raised an exception')
+    except MultipleInvalid as e:
+        assert_equal(2, len(e.errors))
 
 
 # fool-proof test: check that applying `to_dto` and then `to_native` methods